1008번 : A/B (실수형)

+++ AdevideB_1008.java
@@ -2,10 +2,11 @@ import java.util.Scanner;
 
 public class AdevideB_1008 {
     public static void main(String[] args) {
-        int a, b;
+        double a, b;
         Scanner s = new Scanner(System.in);
-        a = s.nextInt();
-        b = s.nextInt();
+        // 정수형으로 계산할 시 소수점 이하를 버리기 때문에 실수형으로 계산
+        a = s.nextDouble();
+        b = s.nextDouble();
         System.out.print(a/b);
     }
 }
